chore: rename `queue.py` -> `celeriac_queue.py`
Rationale: packages/solution will use `queue` from `python`. Remove dependency/import issues
