+++ test_celeriac.py
@@ -1,5 +1,5 @@
 import time
-from queue import Celeriac
+from celeriac_queue import Celeriac
 from unittest.mock import patch
 
 # Import the functions from main.py
